requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0
timm>=0.9.0

# OCR and Document Processing
pytesseract>=0.3.10
opencv-python>=4.8.0
pillow>=10.0.0
pdf2image>=1.16.0
pdfplumber>=0.9.0
pdfminer.six>=20221105
python-docx>=0.8.11
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Computer Vision and Layout Analysis
detectron2 @ git+https://github.com/facebookresearch/detectron2.git
layoutparser>=0.3.4
easyocr>=1.7.0

# API and Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0

# Task Queue and Caching
celery>=5.3.0
redis>=4.5.0
flower>=2.0.0

# Database and Storage
sqlalchemy>=2.0.0
alembic>=1.11.0
boto3>=1.28.0
minio>=7.1.0

# Streamlit and UI
streamlit>=1.25.0
plotly>=5.15.0
altair>=5.0.0
streamlit-aggrid>=0.3.4
streamlit-option-menu>=0.3.6

# Data Processing and Analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Monitoring and Logging
prometheus-client>=0.17.0
structlog>=23.1.0
rich>=13.4.0

# Configuration and Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0
tqdm>=4.65.0
requests>=2.31.0

# Training and Optimization
optuna>=3.2.0
wandb>=0.15.0
tensorboard>=2.13.0

# Model Export and Optimization
onnx>=1.14.0
onnxruntime>=1.15.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.24.0

# Development Tools
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.4.0

# Security and Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Image Processing and Augmentation
albumentations>=1.3.0
imgaug>=0.4.0

# Distributed Training
deepspeed>=0.9.0

# Optional Cloud Services (disabled by default)
# google-cloud-vision>=3.4.0
# azure-cognitiveservices-vision-computervision>=0.9.0
# aws-textract-caller>=0.0.27
